Test green after template reorg
