+++ MetX/MetX.Aspects/GenerateAddingStaticCode.cs
@@ -7,15 +7,4 @@ namespace MetX.Aspects
     {
         
     }
-    
-    [AttributeUsage(AttributeTargets.Class)]
-    public class GenerateFromTemplate : Attribute 
-    {
-        public readonly string TemplateFilePath;
-
-        public GenerateFromTemplate(string templateFilePath)
-        {
-            TemplateFilePath = templateFilePath;
-        }
-    }
 }
