+++ MetX/MetX.Controls/HighlightGroup.cs
@@ -1,0 +1,37 @@
+﻿using System;
+using System.Collections.Generic;
+using ICSharpCode.TextEditor;
+using ICSharpCode.TextEditor.Document;
+
+namespace MetX.Controls
+{
+    /// <summary>Bundles a group of markers together so that they can be cleared 
+    /// together.</summary>
+    public class HighlightGroup : IDisposable
+    {
+        List<TextMarker> _markers = new();
+        TextEditorControl _editor;
+        IDocument _document;
+        public HighlightGroup(TextEditorControl editor)
+        {
+            _editor = editor;
+            _document = editor.Document;
+        }
+        public void AddMarker(TextMarker marker)
+        {
+            _markers.Add(marker);
+            _document.MarkerStrategy.AddMarker(marker);
+        }
+        public void ClearMarkers()
+        {
+            foreach (var m in _markers)
+                _document.MarkerStrategy.RemoveMarker(m);
+            _markers.Clear();
+            _editor.Refresh();
+        }
+        public void Dispose() { ClearMarkers(); GC.SuppressFinalize(this); }
+        ~HighlightGroup() { Dispose(); }
+
+        public IList<TextMarker> Markers => _markers.AsReadOnly();
+    }
+}
