+++ MetX/MetX.Controls/Properties/Resources.cs
