+++ MetX/MetX.Generators.Samples.Client/EmptyClass.cs
@@ -1,0 +1,5 @@
+﻿namespace MetX.Generators.Samples.Client
+{
+    [MetX.Aspects.GenerateAddingStaticCode]
+    public class EmptyClass { }
+}
