+++ MetX/MetX.Generators.Samples.Client/Fred.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace MetX.Generators.Samples.Client
+{
+    [Aspects.GenerateFromTemplate("exampleOne.template")]
+    public class Fred
+    {
+        public string George;
+        public long Frank;
+        public DateTime Mary;
+    }
+}
