+++ MetX/MetX.Generators.Samples.Client/Program.cs
@@ -11,15 +11,4 @@ namespace MetX.Generators.Samples.Client
             Generated.FredExampleOne.Go();
         }
     }
-
-    [MetX.Aspects.GenerateAddingStaticCode]
-    public class EmptyClass { }
-    
-    [Aspects.GenerateFromTemplate("exampleOne.template")]
-    public class Fred
-    {
-        public string George;
-        public long Frank;
-        public DateTime Mary;
-    }
 }
