+++ MetX/MetX.Generators/Templates/CSharp/Default/Namespace.AspectsName/GenerateFromTemplate.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace ~~Namespace~~.~~AspectsName~~
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public class GenerateFromTemplateAttribute : Attribute 
+    {
+        public readonly string TemplateFilePath;
+
+        public GenerateFromTemplate(string templateFilePath)
+        {
+            TemplateFilePath = templateFilePath;
+        }
+    }
+}
