+++ MetX/MetX.Generators/Templates/CSharp/Default/Namespace.AspectsName/GenerateGeneratorName.cs
@@ -1,0 +1,10 @@
+﻿~~:Reference MetX.Windows.dll
+using System;
+
+namespace ~~Namespace~~.~~AspectsName~~
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public class Generate~~GeneratorName~~Attribute : Attribute 
+    {
+    }
+}
