+++ MetX/MetX.Generators/Templates/CSharp/Default/Namespace.AspectsName/Namespace.AspectsName.csproj
