+++ MetX/MetX.Generators/Templates/CSharp/Default/Namespace.ClientName/Namespace.ClientName.csproj
