+++ MetX/MetX.Generators/Templates/CSharp/Default/Namespace.GeneratorsName/Namespace.GeneratorsName.csproj
