+++ MetX/MetX.Glove.Console/Properties/Resources.cs
