+++ MetX/MetX.QuickScripts/MetX.QuickScripts.csproj
@@ -17,6 +17,12 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Templates\Single file input\Exe.cs" />
+    <Compile Remove="Templates\Single file input\Native.cs" />
+    <Compile Remove="Templates\Single file input\qscriptd3cdf9.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\MetX.Aspects\MetX.Aspects.csproj" />
     <ProjectReference Include="..\MetX.Controls\MetX.Controls.csproj">
     </ProjectReference>
@@ -42,6 +48,15 @@
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Templates\Single file input\Exe.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Include="Templates\Single file input\Native.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Include="Templates\Single file input\qscriptd3cdf9.cs" />
   </ItemGroup>
 
 </Project>
