+++ MetX/MetX.QuickScripts/Properties/Settings.cs
