+++ MetX/MetX.QuickScripts/Templates/Single file input/Exe.cs
