+++ MetX/MetX.QuickScripts/Templates/Single file input/Native.cs
