+++ MetX/MetX.Standard/Generation/CSharp/Project/IRefer.cs
@@ -1,0 +1,11 @@
+﻿using System.Xml;
+
+namespace MetX.Standard.Generation.CSharp.Project
+{
+    public interface IRefer
+    {
+        ClientCsProjGenerator Parent { get; set; }
+        XmlElement Remove();
+        XmlElement InsertOrUpdate();
+    }
+}
