+++ MetX/MetX.Standard/MetX.Standard.csproj
@@ -36,9 +36,6 @@
     <Compile Remove="IO\Element.cs" />
     <Compile Remove="Library\ExampleXslUrn.cs" />
     <Compile Remove="Scripts\NArrangeTestLogger.cs" />
-    <Compile Remove="Templates\Single file input\Exe.cs" />
-    <Compile Remove="Templates\Single file input\Native.cs" />
-    <Compile Remove="Templates\Single file input\qscriptd3cdf9.cs" />
     <Compile Remove="Web\HttpAppErrHandler.cs" />
     <Compile Remove="Web\SecureHandler.cs" />
     <Compile Remove="Web\SecurePage.cs" />
@@ -64,23 +61,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Resource Include="Templates\CSharp\Project\ClientFromScratch.csproj">
-      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-    </Resource>
-    <None Include="Templates\CSharp\A\Namespace.AspectsName\Namespace.AspectsName.csproj" />
-    <None Include="Templates\CSharp\A\Namespace.GeneratorsName\Namespace.GeneratorsName.csproj">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-    <None Include="Templates\CSharp\A\Namespace.ClientName\Namespace.ClientName.csproj">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-    <None Include="Templates\Single file input\Exe.cs">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-    <None Include="Templates\Single file input\Native.cs">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-    <None Include="Templates\Single file input\qscriptd3cdf9.cs" />
     <None Include="XLG.snk" />
   </ItemGroup>
   <ItemGroup>
