--- MetX/MetX.Standard/Templates/CSharp/A/Namespace.AspectsName/GenerateGeneratorName.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace ~~Namespace~~.~~AspectsName~~
-{
-    [AttributeUsage(AttributeTargets.Class)]
-    public class Generate~~GeneratorName~~ : Attribute 
-    {
-    }
-}
