+++ MetX/MetX.Tests/Standard/Generation/CSharp/Project/ModifierTests_FromScratch.cs
@@ -1,4 +1,6 @@
-﻿using MetX.Aspects;
+﻿using System;
+using System.IO;
+using MetX.Aspects;
 using MetX.Standard.Generation.CSharp.Project;
 using MetX.Tests.Standard.Generation.CSharp.Project.Pieces;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -11,28 +13,16 @@ namespace MetX.Tests.Standard.Generation.CSharp.Project
         [TestMethod]
         public void ClientFromScratchXmlIsAsExpected()
         {
-            ClientCsProjGenerator clientCsProjGenerator = ClientCsProjGenerator.FromScratch(GenGenOptions.Defaults);
+            var genGenOptions = GenGenOptions.Defaults;
+            genGenOptions.BaseOutputPath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Product");
+            genGenOptions.TemplatesRootPath = @"..\..\..\..\MetX.Generators\Templates";
+            
+            ClientCsProjGenerator clientCsProjGenerator = ClientCsProjGenerator.FromScratch(genGenOptions);
 
             Assert.IsNotNull(clientCsProjGenerator);
             var actual = clientCsProjGenerator.Document.OuterXml;
             Assert.IsFalse(actual.Contains("~~"));
             Assert.IsTrue(actual.Contains("net-5.0windows"));
         }
-
-        [TestMethod]
-        public void TESTNAME_Simple()
-        {
-            var data = new ClientCsProjGenerator(options)
-            {
-                
-            };
-
-            var expected = "EXPECTED";
-            var actual = data.METHODNAME();
-
-            Assert.IsNotNull(actual);
-            Assert.AreEqual(expected, actual);
-        }
-
     }
 }
