+++ MetX/MetX.Tests/Standard/Generation/CSharp/Project/ModifierTests_PackageReference.cs
