+++ MetX/MetX.Tests/Standard/Generation/CSharp/Project/Pieces/Area.cs
@@ -1,0 +1,7 @@
+﻿namespace MetX.Tests.Standard.Generation.CSharp.Project.Pieces
+{
+    public static class Area
+    {
+
+    }
+}
