+++ MetX/MetX.Windows/WinApi/User32.cs
@@ -1,0 +1,30 @@
+﻿using System;
+using System.Runtime.InteropServices;
+
+namespace MetX.Windows.WinApi
+{
+    public class User32
+    {
+        public const short SWP_NOMOVE = 0X2;
+        public const short SWP_NOSIZE = 1;
+        public const short SWP_NOZORDER = 0X4;
+        public const int SWP_SHOWWINDOW = 0x0040;
+
+
+        [DllImport("user32.dll", EntryPoint = "SetWindowPos")]
+        public static extern IntPtr SetWindowPos(IntPtr hWnd, int hWndInsertAfter, int x, int Y, int cx, int cy,
+            int wFlags);
+
+        [DllImport("user32.dll", SetLastError = true)]
+        public static extern bool GetWindowRect(IntPtr hwnd, out RECT lpRect);
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct RECT
+        {
+            public int Left; // x position of upper-left corner
+            public int Top; // y position of upper-left corner
+            public int Right; // x position of lower-right corner
+            public int Bottom; // y position of lower-right corner
+        }
+    }
+}
